eslint rules enforcing dto and entity usage
