+++ eslint_rules/src/rules/entity-use.test.ts
@@ -1,5 +1,5 @@
 import { rule } from "./entity-use.js";
-import { createRuleTester } from "../utils.js";
+import { createRuleTester } from "../test-utils.js";
 
 const ruleTester = createRuleTester();
 
@@ -8,6 +8,10 @@ ruleTester.run("entity-use", rule, {
     {
       filename: "foo.entity.ts",
       code: `let x: FooEntity = y;`,
+    },
+    {
+      filename: "foo.controller.ts",
+      code: `let em:EntityManager;`,
     },
   ],
   invalid: [
